[expo-updates][ios] Refactor AppController into three controllers
