+++ packages/expo-updates/ios/EXUpdates/ReactDelegateHandler/ExpoUpdatesReactDelegateHandler.swift
@@ -17,6 +17,8 @@ public final class ExpoUpdatesReactDelegateHandler: ExpoReactDelegateHandler, Ap
   private var rootViewModuleName: String?
   private var rootViewInitialProperties: [AnyHashable: Any]?
   private lazy var shouldEnableAutoSetup: Bool = {
+    return false
+
     if EXAppDefines.APP_DEBUG && !UpdatesUtils.isNativeDebuggingEnabled() {
       return false
     }
@@ -49,6 +51,7 @@ public final class ExpoUpdatesReactDelegateHandler: ExpoReactDelegateHandler, Ap
     }
 
     self.reactDelegate = reactDelegate
+    AppController.initializeWithoutStarting(configuration: nil)
     let controller = AppController.sharedInstance
     controller.delegate = self
 
@@ -76,7 +79,7 @@ public final class ExpoUpdatesReactDelegateHandler: ExpoReactDelegateHandler, Ap
 
   // MARK: AppControllerDelegate implementations
 
-  public func appController(_ appController: AppController, didStartWithSuccess success: Bool) {
+  public func appController(_ appController: AppControllerInterface, didStartWithSuccess success: Bool) {
     guard let reactDelegate = self.reactDelegate else {
       fatalError("`reactDelegate` should not be nil")
     }
